requirements.txt
Django==4.0.6
django-environ==0.9.0
djangorestframework==3.13.1
environ==1.0
psycopg2==2.9.3